Replace download with content library page
