+++ app/assets/scripts/main.js
@@ -10,7 +10,6 @@ import history from './utils/history'
 
 import Home from './views/home'
 import StaticPage from './views/static-page'
-import Download from './views/download'
 import Library from './views/library'
 import Policies from './views/policies-hub'
 import Policy from './views/policies-page'
@@ -25,7 +24,6 @@ const Root = () => (
       <Switch>
         <Route exact path='/' component={Home}/>
         <Route exact path='/results' component={Results}/>
-        <Route exact path='/download' component={Download}/>
         <Route exact path='/library' component={Library}/>
         <Route path='/compare/:country*' component={Compare}/>
         <Route exact path='/policies/:policyId' component={Policy}/>
